humidity_sensor/requirements.txt
python-nmap
python-dotenv
requests
influxdb_client